Delete old presenter
